Change sizeof to doubles so trialswap doesn't crash
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/picante@146 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ branches/gsoc/src/picante.c
@@ -15,9 +15,9 @@ double **vectomat(double *v,int row,int column) {
   int i,j;
   double **m;
 
-  m = (double **)R_alloc(row,sizeof(int *));
+  m = (double **)R_alloc(row,sizeof(double *));
   for (i=0; i<row; i++) {
-    m[i] = (double *)R_alloc(column,sizeof(int));
+    m[i] = (double *)R_alloc(column,sizeof(double));
     for (j=0; j<column; j++) {
       m[i][j] = v[row*j+i];  /* R uses column-first ordering */
     }
